uigen: add pseudo icon properties
For themed icons, it follows the QQC2 naming convention as "theme" sounds
like a name of the theme itself, and QIcon has name() function.

For the other state pixmaps, it just makes 1:1 mapping to uic nodes. This
looks a bit ugly, but I can't come up with a better and simpler solution.