uigen: remove TODOs about value type naming
I decided to respect C++/Qt naming and behavior than QML/JS. This is the
tool for C++/Qt programmers. We can't disregard the JS/TS syntax for better
editor support (namely Qt Creator), but other than that, we don't need to
import any JavaScript mess. I hate JS with a passion.

It's also simpler to not simulate JS behavior.